+ Optimized imports for app_data.py, main.py + Model__PositionsGrouping.py grouping issue fixed for proxy data + Added freeze support to main.py + Added logging in Model OMS, and updated process to check the checklist + OrderManagerTable.py updated type hints + TradingSymbolMapping.py updated 'ATR TS Multiplier' data type as float (astype usage)
*Multi-accounts order placement working (except iifl)

+++ Libs/Files/TradingSymbolMapping.py
@@ -9,7 +9,7 @@ class StrategiesColumn:
     tradexcb_numeric_columns = {
         'buy_ltp_percent': float, 'sell_ltp_percent': float, 'wait_time': float, 'buy_above': float,
         'sell_below': float, 'stoploss': float, 'tsl': float, 'target': float, 'timeframe': int,
-        'ATR TS Period': int, 'ATR TS Multiplier': int, 'moving_average_period': int}
+        'ATR TS Period': int, 'ATR TS Multiplier': float, 'moving_average_period': int}
 
     strategy_dict = {
         'Default': tradexcb_display_columns,
